Fixes auth issues in Dataset handler
Issue: documentacao-e-tarefas/scielo#941

+++ api/v1/datasets/DatasetHandler.inc.php
@@ -12,11 +12,12 @@ import('plugins.generic.dataverse.classes.services.DataverseService');
 
 class DatasetHandler extends APIHandler
 {
+    private const MANAGER_ROLES = [ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR];
+
     public function __construct()
     {
         $this->_handlerPath = 'datasets';
         $roles = [ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR, ROLE_ID_AUTHOR];
-        $managerRoles = [ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR];
 
         $this->_endpoints = array(
             'GET' => array(
@@ -55,7 +56,7 @@ class DatasetHandler extends APIHandler
                 array(
                     'pattern' => $this->getEndpointPattern() . '/associate',
                     'handler' => [$this, 'associateDataset'],
-                    'roles' => $roles
+                    'roles' => self::MANAGER_ROLES
                 ),
                 array(
                     'pattern' => $this->getEndpointPattern() . '/{studyId}/file',
@@ -72,12 +73,12 @@ class DatasetHandler extends APIHandler
                 array(
                     'pattern' => $this->getEndpointPattern() . '/{studyId}/disassociate',
                     'handler' => [$this, 'disassociateDataset'],
-                    'roles' => $managerRoles
+                    'roles' => self::MANAGER_ROLES
                 ),
                 array(
                     'pattern' => $this->getEndpointPattern() . '/{studyId}/publish',
                     'handler' => array($this, 'publishDataset'),
-                    'roles' => $roles
+                    'roles' => self::MANAGER_ROLES
                 ),
             ),
             'DELETE' => array(
@@ -110,6 +111,32 @@ class DatasetHandler extends APIHandler
         return parent::authorize($request, $args, $roleAssignments);
     }
 
+    private function userHasManagerRole(): bool
+    {
+        $request = Application::get()->getRequest();
+        $user = $request->getUser();
+        $context = $request->getContext();
+
+        if (!$user || !$context) {
+            return false;
+        }
+
+        $roleDao = DAORegistry::getDAO('RoleDAO');
+        return (bool) $roleDao->userHasRole($context->getId(), $user->getId(), self::MANAGER_ROLES);
+    }
+
+    private function userIsAssignedToSubmission(int $submissionId): bool
+    {
+        $user = Application::get()->getRequest()->getUser();
+
+        if (!$user) {
+            return false;
+        }
+
+        $stageAssignmentDao = DAORegistry::getDAO('StageAssignmentDAO');
+        return (bool) $stageAssignmentDao->getBySubmissionAndUserIdAndStageId($submissionId, $user->getId())->next();
+    }
+
     private function mapServiceStatusToHttpCode(string $status): ?int
     {
         $map = [
@@ -127,6 +154,11 @@ class DatasetHandler extends APIHandler
 
         if (!$study) {
             return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
+        }
+
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
         }
 
         try {
@@ -156,6 +188,11 @@ class DatasetHandler extends APIHandler
 
         if (!$study) {
             return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
+        }
+
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
         }
 
         $data = [];
@@ -257,6 +294,10 @@ class DatasetHandler extends APIHandler
         $queryParams = $slimRequest->getQueryParams();
 
         $submissionId = $queryParams['submissionId'];
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
+        }
+
         $draftDatasetFiles = DAORegistry::getDAO('DraftDatasetFileDAO')->getBySubmissionId($submissionId);
 
         if (empty($draftDatasetFiles)) {
@@ -322,6 +363,11 @@ class DatasetHandler extends APIHandler
         $dataverseStudyDAO = DAORegistry::getDAO('DataverseStudyDAO');
         $study = $dataverseStudyDAO->getStudy((int) $args['studyId']);
 
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
+        }
+
         try {
             $dataverseClient = new DataverseClient();
             $datasetFiles = $dataverseClient->getDatasetFileActions()->getByDatasetId($study->getPersistentId());
@@ -348,6 +394,11 @@ class DatasetHandler extends APIHandler
 
         if (!$study) {
             return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
+        }
+
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
         }
 
         try {
@@ -398,6 +449,11 @@ class DatasetHandler extends APIHandler
             return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
         }
 
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
+        }
+
         $datasetFileService = new DatasetFileService();
         $datasetFileService->delete($study, $queryParams['fileId'], $queryParams['filename']);
 
@@ -408,16 +464,21 @@ class DatasetHandler extends APIHandler
     {
         $dataverseStudyDAO = DAORegistry::getDAO('DataverseStudyDAO');
         $study = $dataverseStudyDAO->getStudy((int) $args['studyId']);
-        $deleteMessage = null;
 
+        if (!$study) {
+            return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
+        }
+
+        $submissionId = $study->getSubmissionId();
+        if (!$this->userHasManagerRole() && !$this->userIsAssignedToSubmission($submissionId)) {
+            return $response->withStatus(403)->withJsonError('api.403.unauthorized');
+        }
+
+        $deleteMessage = null;
         $requestParams = $slimRequest->getParsedBody();
         $sendDeleteEmail = (int) $requestParams['sendDeleteEmail'];
         if ($sendDeleteEmail == 1 && isset($requestParams['deleteMessage'])) {
             $deleteMessage = $requestParams['deleteMessage'];
-        }
-
-        if (!$study) {
-            return $response->withStatus(404)->withJsonError('api.404.resourceNotFound');
         }
 
         $datasetService = new DatasetService();
